backend/requirements.txt
APScheduler==3.11.0
Authlib==1.4.0
blinker==1.8.2
boto3==1.36.11
botocore==1.36.11
CacheControl==0.14.1
cachelib==0.13.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
filelock==3.17.0
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
msgpack==1.1.0
msgspec==0.18.6
numpy==2.2.2
oauthlib==3.2.2
opencv-python==4.11.0.86
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygame==2.6.1
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.2
six==1.17.0
smmap==5.0.2
termcolor==2.5.0
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.1.2
